Rust coin state (#12934)
* use rust types for CoinState and RespondToPhUpdates

* improve test_network_protocol_test to support native types

+++ chia/wallet/util/wallet_sync_utils.py
@@ -296,9 +296,9 @@ def get_block_challenge(
 
 def last_change_height_cs(cs: CoinState) -> uint32:
     if cs.spent_height is not None:
-        return cs.spent_height
+        return uint32(cs.spent_height)
     if cs.created_height is not None:
-        return cs.created_height
+        return uint32(cs.created_height)
 
     # Reorgs should be processed at the beginning
     return uint32(0)
